requirements.txt
Flask==3.0.3
tensorflow-cpu==2.20.0
Pillow==10.4.0
gunicorn==23.0.0
